Post Meeting 11/8/13
+ Added a processing program that reads the button and joystick values
from the controller (Controller_Printout).

+ Modified communication protocol used in SerialMotorControl and
DocksideController to include a start bit to fix the LED board.

Great meeting all around! We got the LED board working perfectly on all
4 LEDs (and 4 direction LEDs), and moved on to a (successful!) test
with a single motor. The motor worked well both in air and water, and
drew significantly less current than expected (<1 amp). Additionally,
we figured out that if you have the motor on full power underwater, and
then suddenly reverse direction, you can get a pretty decent plume of
water to splash across the lab. Even better, the motor controller did
not blow up this time (due to both new constraints in the programming
and properly setting the current limiter on the benchtop power supply).

-Charles