Changed the code to work with new shorter message
The methods used build on the ones from Methods/HelpfulMethods.ino.
Testing needed: There may be some off-by-one errors and we may need to
change the processing code for sending servo values because it sends
them in range [1,180] instead of [2,181] (to allow for the start
delimiter char). The code compiles